main: split the code in multiple files
