+++ Makefile.am
@@ -11,7 +11,7 @@ CLEANFILES = fuse-overlayfs.1
 fuse_overlayfs_CFLAGS = -I . -I $(abs_srcdir)/lib $(FUSE_CFLAGS)
 fuse_overlayfs_LDFLAGS =
 fuse_overlayfs_LDADD = lib/libgnu.a $(FUSE_LIBS)
-fuse_overlayfs_SOURCES = main.c
+fuse_overlayfs_SOURCES = main.c direct.c utils.c
 
 WD := $(shell pwd)
